Removed old project Remake new project Added UI
