+++ Archive-Extraction-tool/Archive-Extraction-tool.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31424.327
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Archive-Extraction-tool", "Archive-Extraction-tool\Archive-Extraction-tool.csproj", "{27B023E4-60CF-40A3-9459-F4CB1AB19F5C}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{27B023E4-60CF-40A3-9459-F4CB1AB19F5C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{27B023E4-60CF-40A3-9459-F4CB1AB19F5C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{27B023E4-60CF-40A3-9459-F4CB1AB19F5C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{27B023E4-60CF-40A3-9459-F4CB1AB19F5C}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {94356F90-6473-429C-AF96-91D9D20B1501}
+	EndGlobalSection
+EndGlobal
