--- Unzipper/Unzipper.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30804.86
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Unzipper", "Unzipper\Unzipper.csproj", "{5BF9F817-B0CA-46FB-BDC5-65769E02FCEA}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{5BF9F817-B0CA-46FB-BDC5-65769E02FCEA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5BF9F817-B0CA-46FB-BDC5-65769E02FCEA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5BF9F817-B0CA-46FB-BDC5-65769E02FCEA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5BF9F817-B0CA-46FB-BDC5-65769E02FCEA}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {127DE923-C333-4F8C-86BD-3D149704B3DD}
-	EndGlobalSection
-EndGlobal
