--- Unzipper/Unzipper/MainWindow.xaml
@@ -1,19 +1,0 @@
-﻿<Window x:Class="Unzipper.MainWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:Unzipper"
-        mc:Ignorable="d"
-        Title="Unzipper" Height="450" Width="800" Background="Gray">
-    <Grid>
-        <TextBox x:Name="txtBrowse" HorizontalAlignment="Left" Height="23" Margin="68,71,0,0" TextWrapping="Wrap" VerticalAlignment="Top" Width="151"/>
-        <Button x:Name="BtnBrowse" Content="Browse ..." HorizontalAlignment="Left" Margin="241,71,0,0" VerticalAlignment="Top" Width="79" Height="23" Click="BtnBrowse_Click"/>
-        <TextBox x:Name="txtReport" Text="{Binding Text}" HorizontalAlignment="Left" Height="101" Margin="68,295,0,0" TextWrapping="Wrap" VerticalAlignment="Top" Width="646" TextChanged="txtReport_TextChanged"/>
-        <Button x:Name="BtnClear" Content="Clear" HorizontalAlignment="Left" Margin="68,270,0,0" VerticalAlignment="Top" Width="75" Click="BtnClear_Click"/>
-        <ProgressBar x:Name="Prg_progress" HorizontalAlignment="Left" Height="24" Margin="68,227,0,0" VerticalAlignment="Top" Width="646" ValueChanged="Prg_progress_ValueChanged"/>
-        <Label x:Name="lblProgress" Content="{Binding CurrentProgress}" HorizontalAlignment="Left" Margin="68,196,0,0" VerticalAlignment="Top"/>
-        <Label x:Name="lblEstimated" Content="{Binding Progress}" HorizontalAlignment="Left" Margin="585,196,0,0" VerticalAlignment="Top"/>
-
-    </Grid>
-</Window>
