--- Unzipper/Unzipper/packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Ookii.Dialogs.Wpf" version="3.1.0" targetFramework="net48" />
-  <package id="System.IO.Compression.ZipFile" version="4.3.0" targetFramework="net48" />
-</packages>
